refactor web socket into generic component base and childs
